Add an FTP file system and connection dialog
This utilises the native Java FileSystem libraries to make operations
easy. Unfortunately, I also don't truly understand FTP or how to debug
this file system, so I am having difficulty with getting it to work
properly. It works to list files/directories on some public FTP sites,
but not on a local one when testing.