Enable unarchiving of ZIP files
This overrides the default behaviour of "enter" when trying to open a
file in the browser. The implementation is generic in that an open
action can specify strategies in priority order, and that the supported
archive files are able to be added easily using the service-loader
pattern.

N.B., After unarchiving a file, it will not refresh the browser.
Refreshing is not a feature I have got to yet, so to view the changes
you'll need to see them in a separate browser or re-open this one.