Add some unit tests for a variety of classes
Unfortunately this is not a complete an exhaustive test suite. I hope it
shows some of my thought process into how I would like to test some of
the components.

When writing these, I found that some classes were difficult to test due
to the use of static methods (either through Files, or other
services/modules). I would like to learn how to use dependency injection
better (rather than a "poor mans" version of just passing through
factories in constructors) which may help with alleviating some of the
hardcoded class references.

See the ZipArchiveFileTests for an example of how I planned to test
classes that referenced Files. This test covers creating a test zip file
on an in memory file system which is torn down after each example.