Code review feedback - enable inline browsing of archive files
This method utilises the ZipFileSystem that comes with the JDK. The
benefit of this is that it's all supported without the need for any
libraries.

Unfortunately, this file system only supports inline browsing of zips
on the default file system (i.e., it won't work for nested zips or
for zips on an FTP file system). To get around this, in the case where
the zip is not on the default file system we extract it to a temporary
file.