Code review feedback - support previewing all text content types (not just text/plain)
There are many text types that we may wish to preview (such as code).
Restricting to "plain" limits the usefulness of the preview feature.

Further, this commit adds a custom file type detector that utilises Tika
for detection. By default some file systems do not have great file type
detection capabilities, so including this with the application will help
to guarantee the preview feature is available for use.