Performance improvement for FTP file system - don't filter hidden files and perform directory listing in a background thread
A user would generally want to see all files on an FTP file system,
which is beneficial for us as it's quite expensive to call
Files.isHidden for each individual file.

Further, due to the prior use of the Stream interface, the filter was
only being applied lazily, causing the lookup to happen on the EDT and
making the UI hang. Instead, collect all the files into a list before
returning to the EDT.