Code review feedback - move and inline scrolling from TabbedFileBrwoser into FileBrwoser
The use of the scrollpane surrounding the splitpane meant that sometimes
scrolling would be performed for both the preview and the list, rather
than just one or the other. The scrollpane is meant to be defined on
each of them and not the parent.

Inlined the scrollpane logic as the ScrollableView was unnecessary and
would have required extra work to get the JList to work with it.