Use a CachedThreadPool rather than the common ForkJoin pool for background operations
This swap has been implemented in a singleton manner, with the
BowserFrame managing the shutdown of the executor service when the frame
is dismissed.

This change was made as the common fork join pool has a limited number
of threads available and if those get all hung up, then user actions no
longer do anything. I believe this thread pool type is a good choice for
this task as:
1) a user is generally either "heavily using" the browser and may need
multiple threads quickly spun up, or
2) a user is "not using" the browser and it is safe to kill the threads
to tidy up.

The background IO operations are generally very quick (list a
directories files, etc.). They are sometimes longer depending on the
file system, but roughly speaking, the threads that are spun up should
get re-used efficiently between user actions.

I have not used shutdownNow yet as I believe for a file browser this
should be a user-confirmed action, just in case the user is writing or
moving files between locations.