Remove Tika - it was causing issues with browsing FTP file systems
Not fully sure what was going on, but I narrowed down the cause to this
dependency. Sometimes when probing the content type of a file it was
causing an exception on FTP file systems, permanently killing off FTP
clients and rendering it unusable.

I attempted to use just the filename probing feature, however this
caused threads to hang around in the background and the app to not
properly close.