Update sort comparators to be on the FileNode rather than the Path
This allows support for attribute sorting. The main benefit here is that
the FileNode should have all the data pre-loaded, so the sort algorithm
can operate on the GUI thread efficiently and perform no IO operations.

Unfortunately, retieving file attributes for FTP is still very slow
(this seems like a recurring theme), so for now this is disabled,
meaning sorting by file size is also disabled.